pyproject.toml
[tool.poetry]
name = "laptop-recommender"
version = "0.1.0"
description = ""
authors = ["toxicNinja04 <gupta.manish2304@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
pandas = "^2.2.3"
matplotlib = "^3.9.2"
langchain-llm = "^0.4.15"
langchain-openai = "^0.2.10"
python-dotenv = "^1.0.1"
markdown = "^3.7"
markdown2 = "^2.5.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
